requirements.txt
discord.py
flask
ntlk